newTestServer を main_test に移す
